# admin-panel/api-json/generated-source/models/admin.ts
/* tslint:disable */
/* eslint-disable */
export interface Admin {
  apellidos: string;
  dni?: string;
  email: string;
  f_nacimiento?: string;
  nombres: string;
  password: string;
  rol: string;
  telefono?: string;
}