# admin-panel/api-json/generated-source/models/cliente.ts
/* tslint:disable */
/* eslint-disable */
export interface Cliente {
  Dep?: number;
  Identif?: string;
  Loc?: number;
  '_id'?: string;
  ayn: string;
  codp: string;
  codp2?: string;
  codt?: string;
  codt2?: string;
  deuda?: boolean;
  dir: string;
  email: string;
  feci?: string;
  fecu?: string;
  idc: number;
  mark?: number;
  mascotas?: Array<string>;
  motuv?: string;
  obra?: number;
  obs?: string;
  tel: string;
  tel2?: string;
  telC?: string;
  telC2?: string;
}