# back/src/models/cliente.ts
"use strict";
import mongoose from "mongoose";
import { IUser } from "./user";
import { IMascota } from "./mascota";

const Schema = mongoose.Schema;

/**
 * Interfaz de cliente. Necesario para implementacion
 */
export interface ICliente extends IUser{
  _id: string,
  idc: number,
  ayn: string,
  dir: string,
  codp: string,
  codt: string,
  tel: string,
  telC: string,
  email: string,
  obs: string,
  mark: number,
  feci: Date,
  fecu: Date,
  motuv: string,
  deuda: boolean,
  obra: number,
  Loc: number,
  Dep: number,
  tel2: string,
  telC2: string,
  codt2: string,
  codp2: string,
  Identif: string,
  mascotas: mongoose.Schema.Types.ObjectId[];
}

// Creacion de clase cliente a traves de la interfaz 
const Cliente = new Schema<ICliente>({
  _id: { type: String, required: true },
  idc: { type: Number, required: true },
  ayn: { type: String, required: true },
  dir: { type: String, required: true },
  codp: { type: String, required: false },
  codt: { type: String, required: false },
  tel: { type: String, required: false },
  telC: { type: String, required: false },
  email: { type: String, required: false },
  obs: { type: String, required: false },
  mark: { type: Number, required: true },
  feci: { type: Date, required: true },
  fecu: { type: Date, required: true },
  motuv: { type: String, required: false },
  deuda: { type: Boolean, required: true },
  obra: { type: Number, required: true },
  Loc: { type: Number, required: true },
  Dep: { type: Number, required: true },
  tel2: { type: String, required: false },
  telC2: { type: String, required: true },
  codt2: { type: String, required: false },
  codp2: { type: String, required: false },
  Identif: { type: String, required: true },
  mascotas: [{ type: mongoose.Schema.Types.ObjectId, ref: 'mascotas' }],
})

// Exportacion de la clase
module.exports = mongoose.model('clientes', Cliente);