# back/src/openapi.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
  /provincias/save_provincia:
    post:
      summary: Save a province
      operationId: saveProvince
      tags:
        - Provincia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provincia'
      responses:
        '200':
          description: Province saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provincia'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /provincias/get_provincias:
    get:
      summary: Get all provinces
      operationId: findAllPagingProvince
      tags:
        - Provincia
      security:
        - bearerAuth: []
      parameters:
        - name: filtro
          in: query
          description: filtro
          required: true
          schema:
            type: string
        - name: pagina
          in: query
          description: Pagina
          required: true
          schema:
            type: string
        - name: pageSize
          in: query
          description: Estado del cliente
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of provinces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provincia'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /localidades/save_localidad:
    post:
      summary: Save a locality
      operationId: saveLocality
      tags:
        - Localidad
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Localidad'
      responses:
        '200':
          description: Locality saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Localidad'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /localidades/get_localidades:
    get:
      summary: Get all localities
      operationId: findAllPagingLocalities
      tags:
        - Localidad
      security:
        - bearerAuth: []
      parameters:
        - name: filtro
          in: query
          description: filtro
          required: true
          schema:
            type: string
        - name: pagina
          in: query
          description: Pagina
          required: true
          schema:
            type: string
        - name: pageSize
          in: query
          description: Estado del cliente
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of localities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Localidad'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /save_client:
    post:
      summary: Save a client
      operationId: saveClient
      tags:
        - Cliente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Client saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /login_client:
    post:
      summary: Log in as a client
      operationId: loginClient
      tags:
        - Cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /get_clients:
    get:
      summary: Get all clients
      operationId: findAllPagingClients
      tags:
        - Cliente
      security:
        - bearerAuth: []
      parameters:
        - name: filtro
          in: query
          description: filtro
          required: true
          schema:
            type: string
        - name: pagina
          in: query
          description: Pagina
          required: true
          schema:
            type: string
        - name: pageSize
          in: query
          description: Estado del cliente
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /save_admin:
    post:
      summary: Register a new admin
      operationId: saveAdministrator
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Admin'
      responses:
        '200':
          description: Admin registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /login_admin:
    post:
      summary: Admin login
      operationId: loginAdmin
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLogin'
      responses:
        '200':
          description: Admin logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Provincia:
      type: object
      properties:
        _id:
          type: string
        id:
          type: integer
        nom:
          type: string
        Ctel:
          type: string
        ran:
          type: integer
      required:
        - _id
        - id
        - nom
        - Ctel
        - ran
    Localidad:
      type: object
      properties:
        _id:
          type: string
        id:
          type: integer
        nom:
          type: string
        dep:
          type: integer
        cp:
          type: string
        ran:
          type: integer
      required:
        - _id
        - id
        - nom
        - dep
        - cp
        - ran
    Cliente:
      type: object
      properties:
        _id:
          type: string
        idc:
          type: integer
        ayn:
          type: string
        dir:
          type: string
        codp:
          type: string
        codt:
          type: string
        tel:
          type: string
        telC:
          type: string
        email:
          type: string
        obs:
          type: string
        mark:
          type: number
        feci:
          type: string
          format: date-time
        fecu:
          type: string
          format: date-time
        motuv:
          type: string
        deuda:
          type: boolean
        obra:
          type: integer
        Loc:
          type: integer
        Dep:
          type: integer
        tel2:
          type: string
        telC2:
          type: string
        codt2:
          type: string
        codp2:
          type: string
        Identif:
          type: string
        mascotas:
          type: array
          items:
            type: string
      required:
        - idc
        - ayn
        - dir
        - codp
        - tel
        - email
    ClienteLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    Admin:
      type: object
      properties:
        nombres:
          type: string
        apellidos:
          type: string
        email:
          type: string
        password:
          type: string
        telefono:
          type: string
        rol:
          type: string
        f_nacimiento:
          type: string
        dni:
          type: string
      required:
        - nombres
        - apellidos
        - email
        - password
        - rol
    AdminLogin:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    Token:
      type: object
      properties:
        token:
          type: string
  responses:
    UnauthorizedError:
      description: Unauthorized access